Introduce github-pull-request-mirror script
This change introduces the first version of github-pull-request-mirror.
The script can act as the end point to a Github web hook. Upon receiving
notification that a pull request to one of the user's repositories was
made, it creates a mirror of the branch to be merged in in the
repository the pull request was made in.