Explicitly filter CalledProcessError exceptions
In the past we have seen Git invocation failures on certain requests.
While those have been fixed, it was noted that the entire command as it
was invoked was contained in the resulting exception. That can be a
source of secret leakage, as we supply our token directly as an argument
to Git.
To mitigate those problems, we explicitly handle CalledProcessError
exceptions by not sending their content back in the response.