Remove mirrored branch on pull request close
Right now the branches that we created which mirror the pull request's
content just stay around until the user removes then. That is fine as
long as the pull request rate is low. However, we can also very easily
clean up, in case we get a notification that the pull request was
closed.
This change implements this logic.