Only mirror pull requests from a different repository
When the owner of a repository creates a pull request against the same
repository, the web hooks logic will kick in and a mirror of the source
branch will be created. However, in that very case that is unnecessary,
because the owner's push to the repository (which had to happen to be
able to create a pull request to begin with) will have already ensured
that a CI run was triggered.
Hence, we can safely exclude cases where the source and the destination
repository are the same. This change does so.