Add support for "attaching" pdb
We have seen cases where the program seized to respond to incoming
requests, but have not been able to pin down the problem precisely.
Because the issue is not easily reproducible, this change adds
infrastructure to enter the Python debugger by sending SIGUSR1 to the
running program. Hopefully, this will allow us to root cause the problem
once it reoccurs.